done test message_utest

+++ net/message_fixed.h
@@ -7,7 +7,8 @@ extern "C" {
 
 struct Message_out *message_fixed_create(enum MESSAGE_FIXED type, uint8_t *data, unsigned short delay);
 
-uint8_t message_fixed_proccess(uint8_t *data);
+unsigned short message_fixed_proccess(uint8_t *data);
+
 #ifdef __cplusplus
 } // closing brace for extern "C"
 #endif
