+++ net/message_variable.h
@@ -7,7 +7,7 @@ extern "C" {
 
 uint8_t *message_variable_create(enum MESSAGE_FIXED type, unsigned short *data, unsigned short length, unsigned short delay);
 
-uint8_t message_variable_proccess(uint8_t *data);
+unsigned short message_variable_proccess(uint8_t *data);
 #ifdef __cplusplus
 } // closing brace for extern "C"
 #endif
